Prepared proc_file_data module
