+++ studio/app/optinist/core/expdb/expdb_data.py
@@ -104,29 +104,6 @@ class ExpDbPathIdsUtil:
         return ExpDbPathIds(exp_id=exp_id)
 
 
-@dataclass
-class BatchProcFile:
-    command: str
-    roi_method: Optional[str] = None
-
-
-@dataclass
-class BatchProcFileExt(BatchProcFile):
-    exp_id: Optional[str] = None
-    file_path: Optional[str] = None
-
-    def __post_init__(self):
-        if self.exp_id:
-            exp_ids = ExpDbPathIds(exp_id=self.exp_id)
-            self.file_path = join_filepath(
-                [
-                    EXPDB_DIRPATH.EXPDB_DIR,
-                    exp_ids.subject_id,
-                    f"{self.exp_id}.proc",
-                ]
-            )
-
-
 class ProcessResult:
     def __init__(self):
         self.success_ids_ = []
